Gametext should not be remove on spawn?
Hi,
I'm not sure I am 100% right here, but look.
I had problem and I didn't know why my gametext wouldn't show one connection if I enabled this fix.
Then I realized it was because I was spawning the player.
The thing is I used a gametext to show a "Please Wait" message while I do my operation behind a black screen, involving but not limited to, spawning the player.
I don't think we should remove the GameTexts on spawn. I do understand where it's comign from but this is not the default behavior nor a bug, so I think we should keep it that way if there is no other problem I am not aware of.
I think we should not remove the GameTexts on spawn simply because it is not the normal behavior and our goal is not to change anything, just fix bugs. The debate is opened.
Thank you,
rt-2